refactor: ref Search
