+++ src/view/components/SearchSkeleton/__tests__/SearchSkeleton.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import SearchSkeleton from '../SearchSkeleton';
 
-jest.mock('@containers/search/components/Channel/Channel.jsx', () => () => (
+jest.mock('@containers/search/components/Channel/Channel', () => () => (
   <div id="channel-mock">Hola, soy un item de channel</div>
 ));
