+++ src/view/containers/search/__tests__/Search.test.js
@@ -7,10 +7,10 @@ import Search, { loader as searchLoader } from '../Search';
 import { getPodcastChannels } from '@App/services/podcasts';
 import noop from '@utils/noop';
 
-jest.mock('../components/Channel/Channel.jsx', () => (props) => (
+jest.mock('../components/Channel/Channel', () => (props) => (
   <div>Channel{JSON.stringify(props)}</div>
 ));
-jest.mock('../components/SearchBox/SearchBox.jsx', () => (props) => (
+jest.mock('../components/SearchBox/SearchBox', () => (props) => (
   <div>SearchBox {JSON.stringify(props)}</div>
 ));
 
@@ -18,7 +18,7 @@ jest.mock('@components/SearchSkeleton/SearchSkeleton', () => () => (
   <div>SearchSkeleton</div>
 ));
 
-jest.mock('../../../../services/podcasts');
+jest.mock('@App/services/podcasts');
 
 describe('Search unit test', () => {
   it('should match snapshot when isFetching shows skeleton', () => {
