+++ src/view/containers/search/components/Channel/Channel.tsx
