--- src/view/containers/search/components/Channel/index.js
@@ -1,1 +1,0 @@
-export { default } from './Channel';
