--- src/view/containers/search/components/SearchBox/index.js
@@ -1,1 +1,0 @@
-export { default } from './SearchBox';
