--- src/view/containers/search/index.js
@@ -1,1 +1,0 @@
-export { default, loader } from './Search';
